Incorrect log entry during startup in 4.0
Patch by Tommy Stendahl; Reviewed by xxxx for CASSANDRA-14836

+++ src/java/org/apache/cassandra/net/StartupClusterConnectivityChecker.java
@@ -114,7 +114,7 @@ public class StartupClusterConnectivityChecker
                     TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos),
                     connected,
                     peers.size(),
-                    connected / (peers.size()) * 100.0);
+                    ((float)connected / (float)peers.size()) * 100);
         return succeeded;
     }
